# Fix compiling with `fullgraph=True` to improve inference time.