This PR improves performance by resolving issues occurring when using `torch.compile` with `fullgraph=True`.

Other smaller optimisations include:
- Allocation then filling tensors, rather than multiple allocations then merged with a concatenation that again allocates memory. 
- Removing redundant clone/copy operation.

Future optimisations can include chunked inference. Currently in the below code, `pred_C[0] == audio_eos_value` implicitly requires a sync between CPU and GPU, as it output dependant control-flow. The effect of this can be redude by either: 
- Checking only every $n$ steps, where $n$ is the chunk size. Eg: $n=4$ would mean check every 4 steps.
- Skipping the first $n$, where $n$ is the minimum number of expected inference steps. Eg: $n=86$ would mean check only after the first 1s of generated values.

```python
if (not eos_detected and pred_C[0] == audio_eos_value) or dec_step == max_tokens - max_delay_pattern - 1:
    eos_detected = True
    eos_countdown = max_delay_pattern
```


**Before**

```txt
generate step 86: speed=176.460 tokens/s, realtime factor=2.052x
generate step 172: speed=179.140 tokens/s, realtime factor=2.083x
generate step 258: speed=178.068 tokens/s, realtime factor=2.071x
generate step 344: speed=177.850 tokens/s, realtime factor=2.068x
generate step 430: speed=177.793 tokens/s, realtime factor=2.067x
generate step 516: speed=177.898 tokens/s, realtime factor=2.069x
generate step 602: speed=177.893 tokens/s, realtime factor=2.069x
```

**After**

```txt
generate step 86: speed=196.078 tokens/s, realtime factor=2.280x
generate step 172: speed=201.137 tokens/s, realtime factor=2.339x
generate step 258: speed=200.027 tokens/s, realtime factor=2.326x
generate step 344: speed=200.779 tokens/s, realtime factor=2.335x
generate step 430: speed=200.917 tokens/s, realtime factor=2.336x
generate step 516: speed=200.252 tokens/s, realtime factor=2.329x
generate step 602: speed=199.133 tokens/s, realtime factor=2.315x
```

Benchmarks done on a 4090 @ 300W, on the text in `simple.py`

Edit: using FP16 for both.