@GoudaCouda 2025-05-01
Just for more data I ran this on my 3090 and here are my results. I am using bfloat16 on wsl. Seemed to give either the same or slightly better results

Before
```
generate step 86: speed=4.111 tokens/s, realtime factor=0.048x
generate step 172: speed=56.929 tokens/s, realtime factor=0.662x
generate step 258: speed=52.501 tokens/s, realtime factor=0.610x
generate step 344: speed=60.524 tokens/s, realtime factor=0.704x
generate step 430: speed=58.007 tokens/s, realtime factor=0.675x
generate step 516: speed=62.052 tokens/s, realtime factor=0.722x
generate step 602: speed=60.892 tokens/s, realtime factor=0.708x
generate step 688: speed=61.134 tokens/s, realtime factor=0.711x

```

After

```

generate step 86: speed=4.337 tokens/s, realtime factor=0.050x
generate step 172: speed=58.800 tokens/s, realtime factor=0.684x
generate step 258: speed=61.619 tokens/s, realtime factor=0.717x
generate step 344: speed=61.189 tokens/s, realtime factor=0.711x
generate step 430: speed=60.737 tokens/s, realtime factor=0.706x
generate step 516: speed=62.536 tokens/s, realtime factor=0.727x
generate step 602: speed=58.580 tokens/s, realtime factor=0.681x
generate step 688: speed=63.658 tokens/s, realtime factor=0.740x

```