@JakubCzarlinski 2025-05-01
@GoudaCouda hey there - couple things the results you show don't seem to be post warming up. The first iterations is significantly slower? What results do you get after the model has been through a couple of full generations?

Still about a 3% increase in your environment when excluding the first reported value.

I'll boot into Windows to see what sort of results I get there.