@JakubCzarlinski 2025-05-01
@GoudaCouda to add to my previous comment just letting you know that previous issues in this repo have also reported very slow perf on Windows/WSL - sadly that's just the reality of ML/compiler/HPC development on windows. Briefly touched on it in #160.