@tsdocode 2025-05-01
I've just tried your PR using 'simple.py' with an A100 machine, got result:
```
generate step 516: speed=102.529 tokens/s, realtime factor=1.192x
```

Did you see some warning log like this?:
```
skipping cudagraphs due to mutated inputs (36 instances)
skipping cudagraphs due to mutated inputs (36 instances)
skipping cudagraphs due to mutated inputs (36 instances)
skipping cudagraphs due to mutated inputs (36 instances)
skipping cudagraphs due to mutated inputs (36 instances)
skipping cudagraphs due to mutated inputs (36 instances)
skipping cudagraphs due to mutated inputs (36 instances)
skipping cudagraphs due to mutated inputs (36 instances)
skipping cudagraphs due to mutated inputs (36 instances)
skipping cudagraphs due to mutated inputs (36 instances)
skipping cudagraphs due to mutated inputs (36 instances)
```