@tsdocode 2025-05-01
I also tried with a 4090 machine:
```
generate step 688: speed=125.472 tokens/s, realtime factor=1.459x
```