@JakubCzarlinski 2025-05-01
@tsdocode 

> Did you see some warning log like this?:
> 
> ```
> skipping cudagraphs due to mutated inputs (36 instances)
> skipping cudagraphs due to mutated inputs (36 instances)
> skipping cudagraphs due to mutated inputs (36 instances)
> skipping cudagraphs due to mutated inputs (36 instances)
> skipping cudagraphs due to mutated inputs (36 instances)
> skipping cudagraphs due to mutated inputs (36 instances)
> skipping cudagraphs due to mutated inputs (36 instances)
> skipping cudagraphs due to mutated inputs (36 instances)
> skipping cudagraphs due to mutated inputs (36 instances)
> skipping cudagraphs due to mutated inputs (36 instances)
> skipping cudagraphs due to mutated inputs (36 instances)
> ```

Yes during the first run of the model for compilation I also see those `torch.compile` logs. 

**Could you mention your results before/after?**

---

To be more clear about the methodology for benchmarking: 

1. First load the model with compilation and FP16. Seed the env.
2. Run the model once to trigger the torch compiler. At this stage the model will be very slow. Directly below I have included some of the log during this process from my testing in the WSL environment directly above, with the code from this PR. The kernel finetuning logs can be ignored, the cudagraphs logs can be ignored for now but likely resolving these will also improve performance.

```text
...
skipping cudagraphs due to mutated inputs (36 instances)
skipping cudagraphs due to mutated inputs (36 instances)
skipping cudagraphs due to mutated inputs (36 instances)
skipping cudagraphs due to mutated inputs (36 instances)
generate step 344: speed=121.844 tokens/s, realtime factor=1.417x
skipping cudagraphs due to mutated inputs (36 instances)
skipping cudagraphs due to mutated inputs (36 instances)
skipping cudagraphs due to mutated inputs (36 instances)
...
```

3. Now run the model multiple times to warm-up. Below are the logs from the same WSL env, again with the changes from this PR. Note that there are no more compilation logs.

```text
generate step 258: speed=175.026 tokens/s, realtime factor=2.035x
generate step 344: speed=174.371 tokens/s, realtime factor=2.028x
generate step 430: speed=173.127 tokens/s, realtime factor=2.013x
```